imports and updates on tiltpair events
git-svn-id: file:///home/roberto/Git/svn@11059 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 604caeb26e5957056e84d99b4d50f6bc6a14c6f4

+++ trunk/xmipp/java/src/xmipp/particlepicker/tiltpair/gui/TiltPairPickerJFrame.java
@@ -59,7 +59,7 @@ public class TiltPairPickerJFrame extends ParticlePickerJFrame
 	private MicrographPairsTableModel micrographsmd;
 	private UntiltedMicrograph untiltedmic;
 	private JButton colorbt;
-	private double position;
+	private float position;
 	private int index;
 	private JButton resetbt;
 	private JLabel upslb;
@@ -100,8 +100,8 @@ public class TiltPairPickerJFrame extends ParticlePickerJFrame
 		add(micrographpn, WindowUtils.getConstraints(constraints, 0, 2, 3));
 
 		pack();
-		position = 0.95;
-		WindowUtils.setLocation(position, 0.5, this);
+		position = 0.95f;
+		WindowUtils.setLocation(position, 0.5f, this);
 		setVisible(true);
 	}
 	
@@ -267,7 +267,7 @@ public class TiltPairPickerJFrame extends ParticlePickerJFrame
 						}
 					}, // OK button handler
 					null); // no CANCEL button handler
-			WindowUtils.setLocation(position, 0.5, dialog);
+			WindowUtils.setLocation(position, 0.5f, dialog);
 			dialog.setVisible(true);
 		}
 	}
